Extend strncpy(config.GoCentralAddress) to full field length
config.GoCentralAddress is defined in the ini spec in config.h as being of up to RB3E_MAX_DOMAIN chars, but the strncpy routine for that item was only copying RB3E_MAX_CONFIG_LEN; this was 30 chars. The announced Xbox URI had precisely 31 characters. This caused all DNS lookups for this address to fail, and no connection to be made to the GoCentral server.