massively overhaul index.html
This updates the styling of the page to be a bit more modern. The page also has a theme per the running device state of light mode vs dark mode. Add's the ability to use (or hide) a lastfm api key for album art fetching. This isnt a perfect mechanism so you are also able to hide the api key field to not attempt getting album images. The album images are tagged and cached based on the corresponding song tag, so any arbitrary filtering of the song list will continue to match the cached image. filtered searches trigger faster api calls so the user isnt sat looking at a blank image. once search is unfiltered, the task for populating art for the entire library continues running. No longer need to click a search button to trigger a search, can live type and live filter.